Add files via upload
added input handling, checking for bad URLs

evidently this did not get pushed by github desktop even though it said it did so idk. I don't like github desktop. lol

+++ fancier_ytplist.sh
@@ -1,28 +1,61 @@
 #!/bin/bash
 
-youtube-dl -f ‘bestaudio’ -i -o '%(playlist)s/%(playlist_index)s - %(title)s.%(ext)s' ${1}
+# TODO:
+# prompt for desired directory? OR...
+# create a config file that keeps desired directory
+# make this executable from aywhere?
+# regex for YT playlist URLS
+# (?:(?:PL|LL|EC|UU|FL|RD|UL|TL|PU|OLAK5uy_)[0-9A-Za-z-_]{10,}|RDMM)
+# validate proper playlist URLs?
+# I cannot get the above regex to play nicely and idk why lol
 
-youtube-dl -s ${1} >> tempID.txt
 
-dir=$(./getpldir.sh)
-#using "update.sh" as a var seems to play nicer, idk if it's really necessary but...?
+playlist_url=$1
+
+#testing for input; if no input, ask for input
+
+if [ -z "$1" ]
+  then
+    echo "Please enter the playlist URL or playlist ID."
+    read playlist_url
+    echo "Okay, working..."
+  else
+    echo "Okay, working..."
+fi
+
+# running youtube-dl in simulation mode to extract playlist name
+
+youtube-dl -s $playlist_url >> tempID.txt #dump output to text file
+
+dir=$(./getpldir.sh) #this script extracts playlist name from the text file
+
 script="update.sh"
-#have to put the trailing / in the var itself otherwise it complains
+
 plname=$dir #duplicating this var before I add the slash so I can use it to name the m3u file
+# have to put the trailing / in the var itself otherwise it complains
 dir="$dir/"
+# have to put quotes around the variables when called otherwise it complains
+
+# testing for bad URLs
+if [ -z "$plname" ] #if the getpldir script returns nothing, we will quit
+  then
+  echo "Unable to extract playlist name from URL. Please check that your URL is valid."
+  rm tempID.txt
+  exit 1
+fi
+
+rm tempID.txt #removing tempID file for cleanliness
 
 
-rm tempID.txt
-#have to put quotes around the variables when called otherwise it complains
+youtube-dl -f ‘bestaudio’ -i -o '%(playlist)s/%(playlist_index)s - %(title)s.%(ext)s' $playlist_url
 
-#staging for m3u playlist generation
+# staging for m3u playlist generation
 now=$(date "+%Y-%m-%d-%M") #timestamps are cool and good
 cd "$dir" #changing to newly created directory because it's simpler this way
 
 ls | grep -v .sh | grep -v .m3u >> "$now"_"$plname".m3u #ls piped thru inverse grep works better than plain ls I guess? macOS ships with BSD ls, not GNU ls; no inverse filter with BSD ls.
 
-#ls *.webm >> $now.m3u
-#ls *.m4a >> $now.m3u
+# testing for presence of update.sh, asking how to proceed
 
 if test -f "$script"; then
   echo "$script already exists in the target directory."
@@ -31,7 +64,7 @@ if test -f "$script"; then
     case $yn in
       Yes ) rm $script;
       echo "#!/bin/bash" >> "$script";
-      echo "youtube-dl -f ‘bestaudio’ -i -o '%(playlist_index)s - %(title)s.%(ext)s' ${1}" >> "$script";
+      echo "youtube-dl -f ‘bestaudio’ -i -o '%(playlist_index)s - %(title)s.%(ext)s' $playlist_url" >> "$script";
       echo "now=\$(date \"+%Y-%m-%d-%M\")" >> "$script";
       echo "ls | grep -v .sh | grep -v .m3u >> \$now\_\""$plname"\".m3u" >> "$script";
       #so now the update script will also build a new playlist with each run
@@ -47,7 +80,7 @@ if test -f "$script"; then
 else
   echo "#!/bin/bash" >> "$script"
 
-  echo "youtube-dl -f ‘bestaudio’ -i -o '%(playlist_index)s - %(title)s.%(ext)s' ${1}" >> "$script"
+  echo "youtube-dl -f ‘bestaudio’ -i -o '%(playlist_index)s - %(title)s.%(ext)s' $playlist_url" >> "$script"
   echo "now=\$(date \"+%Y-%m-%d-%M\")" >> "$script"
   echo "ls | grep -v .sh | grep -v .m3u >> \$now\_\""$plname"\".m3u" >> "$script"
   #so now the update script will also build a new playlist with each run
@@ -55,7 +88,7 @@ else
   echo "$script has been written in $(pwd). Run $script within $(pwd) to update the files."
 fi
 
-#this is probably sloppy but I didn't go to computer college, sorry :)
+# this is probably sloppy but I didn't go to computer college, sorry :)
 
 #shoutouts:
 # https://stackoverflow.com/questions/226703/how-do-i-prompt-for-yes-no-cancel-input-in-a-linux-shell-script
